[ENGG-2112] fix: rule saving in different group when created from rules table
